Fix crash when selected fiber isn't in snapshot
